Refactor: split state class into multiple file, one class one file.
