+++ Assets/Scripts/State/MonsterState.cs
@@ -1,0 +1,12 @@
+using Actor;
+
+namespace State
+{
+public class MonsterState : BaseState
+{
+    public virtual BaseState Update(ActorBase actor) { return this; }
+    public virtual BaseState FixedUpdate(ActorBase actor) { return this; }
+    public virtual void OnStateStart(ActorBase actor) {}
+}
+
+}
