+++ Assets/Scripts/State/MovableState.cs.meta
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c9b121cfd601f7f9c92f71c6806c4303d0cd21a285bcab88c10ed9b8ed018952
+size 243
