+++ Assets/Scripts/State/State.cs
@@ -1,0 +1,11 @@
+using Actor;
+
+namespace State {
+public interface BaseState
+{
+    public BaseState Update(ActorBase actor);
+    public BaseState FixedUpdate(ActorBase actor);
+    public void OnStateStart(ActorBase actor);
+}
+
+}
